chore: save

+++ libs/modules/page-builder/src/lib/page-trading-simulator/page-trading-simulator-statistics/components/page-trading-simulator-statistics-buttons.component.ts
@@ -16,25 +16,39 @@ import { TradingSimulatorService } from '@mm/trading-simulator/data-access';
     <div class="flex justify-end gap-x-4">
       @switch (simulatorData().state) {
         @case ('live') {
-          <!-- start button -->
-          <button mat-stroked-button type="button" color="accent" (click)="onStart()">
-            <mat-icon iconPositionEnd>play_arrow</mat-icon>
-            Start
-          </button>
+          @if (isAuthUserOwner()) {
+            <!-- start button -->
+            <button mat-stroked-button type="button" color="accent" (click)="onStart()">
+              <mat-icon iconPositionEnd>play_arrow</mat-icon>
+              Start
+            </button>
+          } @else {
+            @if (!isUserJoined()) {
+              <!-- join -->
+              <button (click)="onJoin()" mat-stroked-button type="button" color="accent">Join Simulator</button>
+            } @else {
+              <!-- leave -->
+              <button (click)="onLeave()" mat-stroked-button type="button" color="warn">Leave Simulator</button>
+            }
+          }
         }
         @case ('started') {
-          <!-- finish button -->
-          <button mat-stroked-button type="button" color="warn" (click)="onFinish()">
-            <mat-icon iconPositionEnd>stop</mat-icon>
-            Finish
-          </button>
+          @if (isAuthUserOwner()) {
+            <!-- finish button -->
+            <button mat-stroked-button type="button" color="warn" (click)="onFinish()">
+              <mat-icon iconPositionEnd>stop</mat-icon>
+              Finish
+            </button>
+          }
         }
         @case ('finished') {
-          <!-- delete button -->
-          <button mat-stroked-button type="button" color="warn" (click)="onDelete()">
-            <mat-icon iconPositionEnd>delete</mat-icon>
-            Delete
-          </button>
+          @if (isAuthUserOwner()) {
+            <!-- delete button -->
+            <button mat-stroked-button type="button" color="warn" (click)="onDelete()">
+              <mat-icon iconPositionEnd>delete</mat-icon>
+              Delete
+            </button>
+          }
         }
       }
     </div>
@@ -56,23 +70,34 @@ export class PageTradingSimulatorStatisticsButtonsComponent {
   readonly isAuthUserOwner = computed(
     () => this.simulatorData()?.owner.id === this.authenticationUserStoreService.state.getUserData().id,
   );
+  readonly isUserJoined = computed(() =>
+    this.simulatorData().participants.includes(this.authenticationUserStoreService.state.getUserData().id),
+  );
 
   @Confirmable('Are you sure you want to start the simulator?')
   onStart() {
-    // start the simulator
-    this.tradingSimulatorService.simulatorStateChangeStart(this.simulatorData());
+    try {
+      // start the simulator
+      this.tradingSimulatorService.simulatorStateChangeStart(this.simulatorData());
 
-    // notify user
-    this.dialogServiceUtil.showNotificationBar('Simulator started', 'success');
+      // notify user
+      this.dialogServiceUtil.showNotificationBar('Simulator started', 'success');
+    } catch (error) {
+      this.dialogServiceUtil.handleError(error);
+    }
   }
 
   @Confirmable('Are you sure you want to finish the simulator?')
   onFinish() {
-    // finish the simulator
-    this.tradingSimulatorService.simulatorStateChangeFinish(this.simulatorData());
+    try {
+      // finish the simulator
+      this.tradingSimulatorService.simulatorStateChangeFinish(this.simulatorData());
 
-    // notify user
-    this.dialogServiceUtil.showNotificationBar('Simulator finished', 'success');
+      // notify user
+      this.dialogServiceUtil.showNotificationBar('Simulator finished', 'success');
+    } catch (error) {
+      this.dialogServiceUtil.handleError(error);
+    }
   }
 
   @Confirmable('Are you sure you want to delete the simulator?')
@@ -85,5 +110,34 @@ export class PageTradingSimulatorStatisticsButtonsComponent {
 
     // notify user
     this.dialogServiceUtil.showNotificationBar('Simulator deleted');
+  }
+
+  async onJoin() {
+    const invitationCode =
+      this.simulatorData().invitationCode !== ''
+        ? await this.dialogServiceUtil.showInlineInputDialog({
+            title: 'Add Invitation Code',
+            description: 'Enter the invitation code to join the trading simulator',
+          })
+        : '';
+
+    // allow sending empty string - then simulator is open for everyone
+    if (invitationCode === null || invitationCode === undefined) {
+      return;
+    }
+
+    try {
+      this.dialogServiceUtil.showNotificationBar('You are joining the simulator');
+      await this.tradingSimulatorService.joinSimulator(this.simulatorData(), invitationCode);
+      this.dialogServiceUtil.showNotificationBar('You have joined the simulator', 'success');
+    } catch (error) {
+      this.dialogServiceUtil.handleError(error);
+    }
+  }
+
+  @Confirmable('Please confirm leaving the simulator')
+  onLeave() {
+    this.tradingSimulatorService.leaveSimulator(this.simulatorData());
+    this.dialogServiceUtil.showNotificationBar('You have left the simulator', 'success');
   }
 }
